:hammer: Refactor(game): 게임 내 고정 값 Enum 리팩토링
- value 값 Double 타입으로 수정

+++ backend/src/main/java/com/sehbeomschool/nova/domain/game/domain/Ages.java
@@ -44,7 +44,7 @@ public class Ages extends BaseEntity {
     public static Ages createStartAge(Integer startSalary) {
         return Ages.builder()
             .totalAsset(Long.valueOf(startSalary))
-            .age(FixedValues.START_AGE.getValue())
+            .age(FixedValues.START_AGE.getValue().intValue())
             .build();
     }
